Move services from subfolders
