+++ server/src/main/kotlin/com/itami/workout_flow/route/workout/WorkoutsRoute.kt
@@ -4,7 +4,7 @@ import com.itami.workout_flow.app.plugins.DEFAULT_RATE_LIMIT
 import com.itami.workout_flow.auth.firebase.FIREBASE_AUTH
 import com.itami.workout_flow.auth.firebase.FirebasePrincipal
 import com.itami.workout_flow.routes.WorkoutsRoute
-import com.itami.workout_flow.service.workout.WorkoutService
+import com.itami.workout_flow.service.WorkoutService
 import io.ktor.http.HttpStatusCode
 import io.ktor.server.auth.authenticate
 import io.ktor.server.auth.principal
